Modify MapFactory.java
